Small optimization on the LCD_I2C.uzi library.
Instead of waiting on the ready() function we just wait until the lcd variable. This saves a few bytes on the encoded program but, most importantly, it avoids having to encode the ready() function and the execution should be much faster because we don't have to execute a script call.