Trying to come up with a good way of toggling panel visibility.
Still not good results, though. The idea was to preserve the previous panel position as an array of indices inside the layout. Then, when restoring the panel we would only need to insert it in its previously stored position.
It seems to work (kinda) when we have small panels like the controls or inspector and hiding them doesn't remove the column entirely. However, for bigger panels that occupy an entire column (like the blocks, for instance) it doesn't work because as soon as I hide the panel, the column is no longer there and the preferred position is now invalid.
The code tries to do its best to position the panel using the data it has but it's still a mess. I don't know yet how to fix this...